+++ src/BUILD
@@ -199,8 +199,6 @@ cc_library(
         "logging.cpp",
         "binaryutils.hpp",
         "binaryutils.cpp",
-        "plugin_configuration.hpp",
-        "plugin_configuration.cpp",
     ],
     deps = [
         "@tensorflow_serving//tensorflow_serving/apis:prediction_service_cc_proto",
